Unreverse order of pronunciations in dict
New code is less efficient, but guarantees that the order of
pronunciations in the cmudict file match the order or pronunciations
for each word in the phonetic dictionary. This matters, since a lot
of the shortcuts I'm currently taking involve handling only the first
pronunciation in the dict...and pronunciations in the input file seem
to be ordered by commonality of use. So this should help a lot.